added the /item logic

+++ src/main/java/com/ap/project/httpHandler/FoodHttpHandler.java
@@ -6,16 +6,21 @@ import com.ap.project.dto.FoodDto;
 import com.ap.project.dto.RestaurantDto;
 import com.ap.project.entity.restaurant.Food;
 import com.ap.project.entity.restaurant.Restaurant;
+import com.ap.project.entity.user.Customer;
 import com.ap.project.entity.user.Seller;
 import com.ap.project.entity.user.User;
-import com.google.gson.Gson;
+import com.google.gson.*;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
+import org.hibernate.Hibernate;
 
+import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.List;
 
 import static com.ap.project.httpHandler.SuperHttpHandler.internalServerFailureError;
 import static com.ap.project.httpHandler.SuperHttpHandler.sendSuccessMessage;
@@ -24,7 +29,7 @@ public class FoodHttpHandler implements HttpHandler {
     public void handle(HttpExchange exchange) throws IOException {
         String path = exchange.getRequestURI().getPath();
         String[] parts = path.split("/");
-        Restaurant restaurant = RestaurantDao.getRestaurantById(Integer.parseInt(parts[2]));
+
         String method = exchange.getRequestMethod();
         User user = SuperHttpHandler.getUserByExchange(exchange);
 
@@ -32,44 +37,53 @@ public class FoodHttpHandler implements HttpHandler {
             return;
         }
 
-        if (!(user instanceof Seller)) {
-            exchange.sendResponseHeaders(403, -1);
-            return;
-        }
-
-        if (restaurant == null) {
-            exchange.sendResponseHeaders(404, -1);
-            System.out.println("Restaurant not found");
-            return;
-        }
-
-        if (RestaurantDao.getSellerId(restaurant.getId()) != user.getUserId()) {
-            exchange.sendResponseHeaders(403, -1);
-            System.out.println("Restaurant not owned by user");
-            return;
-        }
-
-        if (parts.length == 4) {
-            if (method.equals("POST"))
-                handleAddFoodItem(exchange, restaurant);
-            else {
-                exchange.sendResponseHeaders(405, -1);
+        if (path.startsWith("/restaurant")) {
+            Restaurant restaurant = RestaurantDao.getRestaurantById(Integer.parseInt(parts[2]));
+            if (restaurant == null) {
+                exchange.sendResponseHeaders(404, -1);
+                System.out.println("Restaurant not found");
+                return;
             }
-        } else if (parts.length == 5) {
 
-            int foodID = Integer.parseInt(parts[4]);
+            if (RestaurantDao.getSellerId(restaurant.getId()) != user.getUserId()) {
+                exchange.sendResponseHeaders(403, -1);
+                System.out.println("Restaurant not owned by user");
+                return;
+            }
 
-            switch (method) {
-                case "PUT":
-                    handleEditFoodItem(exchange, foodID, restaurant);
-                    return;
-                case "DELETE":
-                    handleDeleteFoodItem(exchange, foodID, restaurant);
-                    return;
-                default:
+            if (parts.length == 4) {
+                if (method.equals("POST"))
+                    handleAddFoodItem(exchange, restaurant);
+                else {
                     exchange.sendResponseHeaders(405, -1);
+                }
+            } else if (parts.length == 5) {
+
+                int foodID = Integer.parseInt(parts[4]);
+
+                switch (method) {
+                    case "PUT":
+                        handleEditFoodItem(exchange, foodID, restaurant);
+                        return;
+                    case "DELETE":
+                        handleDeleteFoodItem(exchange, foodID, restaurant);
+                        return;
+                    default:
+                        exchange.sendResponseHeaders(405, -1);
+                }
             }
-        } else {
+        } else if (path.startsWith("/items")) {
+            if (!(user instanceof Customer)) {
+                exchange.sendResponseHeaders(403, -1);
+                return;
+            }
+
+            if (parts.length == 2)
+                handleGetListOfItems(exchange);
+            else
+                handleGetItemDetails(exchange, Integer.parseInt(parts[2]));
+        }
+        else {
             exchange.sendResponseHeaders(404, -1);
             System.out.println("invalid request");
         }
@@ -116,7 +130,7 @@ public class FoodHttpHandler implements HttpHandler {
             InputStreamReader reader = new InputStreamReader(exchange.getRequestBody(), StandardCharsets.UTF_8);
             FoodDto req = new Gson().fromJson(reader, FoodDto.class);
 
-            Food food = FoodItemDao.getFoodByID(foodId);
+            Food food = FoodItemDao.getFoodByID(foodId, exchange);
 
             if (food == null || food.getRestaurant() == null || food.getRestaurant().getId() != restaurant.getId()) {
                 exchange.sendResponseHeaders(404, -1);
@@ -134,7 +148,7 @@ public class FoodHttpHandler implements HttpHandler {
 
     public void handleDeleteFoodItem(HttpExchange exchange, int foodId, Restaurant restaurant) throws IOException {
         try {
-            Food food = FoodItemDao.getFoodByID(foodId);
+            Food food = FoodItemDao.getFoodByID(foodId, exchange);
             if (food == null || food.getRestaurant() == null || food.getRestaurant().getId() != restaurant.getId()) {
                 exchange.sendResponseHeaders(404, -1);
                 return;
@@ -143,7 +157,60 @@ public class FoodHttpHandler implements HttpHandler {
             sendSuccessMessage("Food successfully removed.", exchange);
 
         } catch (Exception e) {
-            e.printStackTrace();
+            internalServerFailureError(e, exchange);
+        }
+    }
+
+    public void handleGetListOfItems(HttpExchange exchange) throws IOException {
+        try {
+            if (!exchange.getRequestMethod().equals("POST")) {
+                exchange.sendResponseHeaders(405, -1);
+                return;
+            }
+            BufferedReader reader = new BufferedReader(new InputStreamReader(exchange.getRequestBody(), "UTF-8"));
+            StringBuilder sb = new StringBuilder();
+            String line;
+            while ((line = reader.readLine()) != null) {
+                sb.append(line);
+            }
+            String requestBody = sb.toString();
+            JsonObject json = JsonParser.parseString(requestBody).getAsJsonObject();
+
+            String search = json.has("search") && !json.get("search").isJsonNull()
+                    ? json.get("search").getAsString()
+                    : null;
+
+            int price = json.has("price") && !json.get("price").isJsonNull()
+                    ? json.get("price").getAsInt()
+                    : -1;
+
+            List<String> keywords = new ArrayList<>();
+            if (json.has("keywords") && json.get("keywords").isJsonArray()) {
+                for (JsonElement element : json.getAsJsonArray("keywords")) {
+                    keywords.add(element.getAsString());
+                }
+            }
+
+            List<FoodDto> results = FoodItemDao.getItemsByFilters(search, price, keywords);
+            sendSuccessMessage(new Gson().toJson(results), exchange);
+
+
+        } catch (Exception e) {
+            internalServerFailureError(e, exchange);
+        }
+    }
+
+    public void handleGetItemDetails(HttpExchange exchange, int foodId) throws IOException {
+        try {
+            if (!exchange.getRequestMethod().equals("GET")) {
+                exchange.sendResponseHeaders(405, -1);
+                return;
+            }
+            Food food = FoodItemDao.getFoodByID(foodId, exchange);
+            FoodDto foodDto = food.getFoodDto();
+            sendSuccessMessage(new Gson().toJson(foodDto), exchange);
+
+        } catch (Exception e) {
             internalServerFailureError(e, exchange);
         }
     }
